squid/vision.py
"""Vision processing module using Molmo for UI element detection."""
import os
from typing import Dict, Any, List
import pyautogui
from PIL import Image
import torch

class ScreenAnalyzer:
    """Handles screen analysis using Molmo vision system."""

    def __init__(self):
        # Use MPS (Metal Performance Shaders) on Apple Silicon if available
        self.device = "mps" if torch.backends.mps.is_available() else "cpu"
        # TODO: Initialize Molmo components
        self.setup_molmo()

    def setup_molmo(self):
        """Initialize Molmo vision system components."""
        # TODO: Implement Molmo initialization
        # This will be implemented once we have access to Molmo
        pass

    async def capture_screen(self) -> Image.Image:
        """Capture current screen state."""
        return pyautogui.screenshot()

    async def analyze_screen(self) -> Dict[str, Any]:
        """Process screenshot through Molmo to detect UI elements."""
        screenshot = await self.capture_screen()

        # TODO: Process through Molmo pipeline
        # Will implement:
        # 1. Image preprocessing
        # 2. Feature extraction
        # 3. UI element detection
        # 4. Coordinate mapping

        return {
            'screenshot': screenshot,
            'elements': [],  # Will contain detected UI elements
            'coordinates': {},  # Will contain element coordinates
        }
